Iterate over batches of 1000 delete markers (#15)

+++ aws-s3-undelete.sh
@@ -7,8 +7,8 @@ Usage: ${NAME} [-h] [-d] [-p PROFILE] [-v VERSION] BUCKET [PREFIX]
 Removes the delete markers from the given BUCKET and optional PREFIX with the optional
 VERSION id. Objects expired at the same time through lifecycle rules share the same 
 version id.
-It's highly recommended that the PREFIX is specified, because a maximum of 1000 can be
-completed at a time.
+It's highly recommended that the PREFIX is specified to reduce the number of files
+processed at a time.
 And specifying the VERSION id will ensure that only the intended deletions are restored.
 Note: this only works on versioned buckets, where the previous versions still exist!
     -d debug mode: display the aws command that would be run, without running it
@@ -62,16 +62,23 @@ fi
 
 TMP=/tmp/${NAME}-$$.json
 
-# get the delete markers
-aws ${PROFILE} s3api list-object-versions --bucket ${BUCKET} ${PREFIX} \
-    --query "{Objects:DeleteMarkers[${VERSION}]|[0:999].{VersionId:VersionId,Key:Key}}" > ${TMP}
+# iterate over batches of 1000 delete markers until done
+while true; do
+    # get the delete markers
+    aws ${PROFILE} s3api list-object-versions --bucket ${BUCKET} ${PREFIX} \
+        --query "{Objects:DeleteMarkers[${VERSION}]|[0:999].{VersionId:VersionId,Key:Key}}" > ${TMP}
+    if [ $(wc -l ${TMP} | awk '{ print $1 }') -gt 3 ]; then
+        # delete them
+        ${DEBUG} aws ${PROFILE} s3api delete-objects --bucket ${BUCKET} --delete file://${TMP}
 
-# delete them
-${DEBUG} aws ${PROFILE} s3api delete-objects --bucket ${BUCKET} --delete file://${TMP}
+        if [[ ${DEBUG} = "echo" ]]; then
+            cat ${TMP}
+        fi
 
-if [[ ${DEBUG} = "echo" ]]; then
-    cat ${TMP}
-fi
-
-# clean up
-rm -f ${TMP}
+        # clean up
+        rm -f ${TMP}
+    else
+        echo Done
+        exit 0
+    fi
+done
